Now building Lua as a dynamic library
This is to allow Lua plugins that were built using a dynamic Lua library
to run on CraftOS-PC without requiring extra libraries. It also
conveniently closes MCJack123/craftos2#142 in the process!

Instead of importing CraftOS-PC-modified functions in the library, we
are instead using new initialization functions to push the modified
versions to the library. The default versions of the functions will
remain in place until then, allowing linking with this Lua without
having to write all of the functions yourself.