Added substring optimization
This adds a new substring type that stores a pointer to an existing
string + an offset and length. string.sub now creates a new substring
object, removing the need to constantly copy substrings into new
string objects. This helps improve the performance of string.sub
drastically. This has about the same safety guarantees as the rope impl.