lua_tostring now converts substr -> string
This makes programs that rely on `lua_tostring`'s output ending in NUL
work again. If a length is specified to `lua_tolstring`, the substring
will not be converted to a string. If doing this, make sure to use the
length as returned from tolstring and not rely on an ending NUL!