Replaced rope & substr allocation with clusters
This allocates ropes/substrings in clusters instead of individually. These
clusters are arrays of entries that are linked together with a bitmap of
allocated entries at the beginning. When an entry is allocated, the code
searches through each cluster to find if there are any slots available,
and if there are any then the entry is reused and reinitialized. To free
an entry, the used bit in the bitmap is simply cleared. During the
finalization stage of GC, the GC iterates through the entire list and
frees all clusters that are completely empty (except for one for speed).
To optimize, the first known free cluster is cached to avoid having to
iterate through known full clusters.