Moved cluster allocators to external code
The garbage collector wasn't working satisfactorily with the cluster allocation
system, so it has been replaced with a generic object allocation callback.
The cluster code is now located in the embedding application (CraftOS-PC).
This should decouple the implementation details of allocation from
the garbage collector itself, and allows extending the technique to more
object types if helpful.